Convert to using IbmIam2AwsSamlScreenScrape
